Imlement @rizar's suggested edits
